LPS-52624 Copy EditCategoryAction, no logic changes
